BigDecimal Format, helper Methods
Use of BigDecimal will prevent Double Overflow error that plagues double
and float
Addition of Leading value helper methods

+++ src/Vector.java
@@ -1,14 +1,15 @@
+import java.math.BigDecimal;
 
 public class Vector {
 	//Vector x is represented by a horizontal Array;
-	public double[] vector;
+	public BigDecimal[] vector;
 	
 	public Vector(int rows) {
-		vector = new double[rows];
+		vector = new BigDecimal[rows];
 	}
 	
-	public Vector(double[] vec) {
-		vector = new double[vec.length];
+	public Vector(BigDecimal[] vec) {
+		vector = new BigDecimal[vec.length];
         for(int row = 0;row<vec.length;row++)
         	vector[row] = vec[row];
 	}
